+++ findContact/hooks/index.js
@@ -1,0 +1,2 @@
+export { useCategories } from './useCategories';
+export { useFetchContacts } from './useFetchContacts';