+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## (IN PROGRESS)
 
+* Find a contact - Implement MCL Next/Previous pagination. Refs UIPFCONT-29.
+
 ## [3.2.0](https://github.com/folio-org/ui-plugin-find-contact/tree/v3.2.0) (2022-07-07)
 [Full Changelog](https://github.com/folio-org/ui-plugin-find-contact/compare/v3.1.0...v3.2.0)